requirements.txt
absl-py
google
google-api-python-client
google.cloud
google-cloud-storage
fastavro
kafka-python
pyfits
pillow
